feat: add mock service provider URL and implement mock RTP sending functionality
