+++ functional-tests/tests/test_callback.py
@@ -1,5 +1,6 @@
 import allure
 import pytest
+
 from api.callback import srtp_callback
 from utils.dataset import generate_rtp_data
